LPS-111428 Extract Navigation component
